FEAT/BUG: Added file method to delete all matching temp files which are not in use. Added FILE_EXISTS status code, and fixed bug where the correct status code for the DirectoryCreate method was not returned.
The new FileDeleteUnusedTempFiles follows similar logic to the FileGetExistingTempFile method, except instead of returning the first matching file, it deleted all matching files. This is used when initialising a data set to ensure that any temporary files left over from any previous improper shutdowns are cleared up. If a file is in use, it will not be deleted.

File and directory creation methods now return a new status to indicate that the file or directory already exists.

The FileCreateDirectory method did not correctly check the errno macro. Instead, it was checking the error variable which would only ever be 0 or -1.