FEAT/OPTIM: Added aligned malloc methods to be used when allocating a resource handle, and determined the compare and swap to use at compile time rather than runtime.
Alignment is neccessary to ensure that the memory is aligned when doing a double width compare and swap. This is usually done anyway, but with some alloc libraries (e.g. dmalloc) this is not the case, and a segfault will occur.
Determining which compare and swap to use in Windows is now done using the `_WIN64` macro whicih is defined when compiled in x64.