BUG: Fixed exception when running reload example in Windows when compiling in 32 bit.
The `InterlockedCompareExchange64` method was used incorrectly in the `INTERLOCK_EXCHANGE_PTR_DW` macro. While the `InterlockedCompareExchange128` method used in 64 bit compilation returns a boolean, the `InterlockedCompareExchange64` method used in 32 bit compilation returns the 64 bit integer stored in the destination before the exchange.
The macro now compares the return value to the 64 bit integer stored in the desination after the exchange, meaning it `INTERLOCK_EXCHANGE_PTR_DW` has the same behaviour in 32 and 64 bit.