Merged PR 7516: BUG/TEST: Handle truncation of pseudoheaders nicely without an exception, and test the scenario.
The previous behaviour when calling `PseudoHeadersAddEvidence` with headers that would result in a pseudo-header longer than the buffer size given to the method, was to set the exception.
This changes follows the same logic as standard headers, and allows the pseudo-header to be used, but will truncate the value to the length of the buffer. The check as to whether the whole pseudo-header has been written can still be carried out by the caller of the `constructPseudoEvidence` method. Instead of checking for the exception state, the caller should check that the value returned by the method is less than or equal to the buffer length. Return value being greater than the buffer length indicates it was truncated.

Related work items: #4979